Revert "remove odering on list API to improve performance"
This reverts commit 54e45a4d4bbfcdc8625f70a3d284a37762ec3149.

Initially I though that this will improve performance, but
having an up and running app and being able to test different
scenarios, I see that this has no effect. I should monitor &
profile to be able to optimize further but as far as this goes,
it's better to have ordering so that list API response is not
confusing.